Ensure that deleting can be performed only on non-null entities

+++ server/src/main/java/server/api/ParticipantController.java
@@ -98,7 +98,7 @@ public class ParticipantController {
      * on "/api/websocket/v1/channel/{eventCode}/participant
      * with WSAction CREATED
      *
-     * @param eventCode The event code
+     * @param eventCode      The event code
      * @param participantDTO The ParticipantDTO instance
      * @return ResponseEntity with ParticipantDTO or HttpStatus.NOT_FOUND if not found
      */
@@ -135,8 +135,8 @@ public class ParticipantController {
     public ResponseEntity<ParticipantDTO> deleteOne(
             @RequestParam("name") String participantName,
             @PathVariable("eventCode") String eventCode
-    ){
-        try{
+    ) {
+        try {
             Participant participant = participantService.getOne(eventCode, participantName);
             ParticipantDTO participantDTO = participantDTOMapper.toDTO(participant);
 
@@ -152,7 +152,7 @@ public class ParticipantController {
                     ));
 
             return new ResponseEntity<>(participantDTO, HttpStatus.OK);
-        } catch (NotFoundInDatabaseException e){
+        } catch (NotFoundInDatabaseException e) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
@@ -190,14 +190,15 @@ public class ParticipantController {
         }
     }
 
-    private void deleteDependants(Participant participant){
-        for (Expense expense : participant.getPaidForExpenses()){
-            expenseController.deleteOne(
-                    expense.getId(),
-                    expense.getPaidBy().getName(),
-                    participant.getEvent().getCode()
-            );
-        }
+    private void deleteDependants(Participant participant) {
+        if (participant != null)
+            for (Expense expense : participant.getPaidForExpenses()) {
+                expenseController.deleteOne(
+                        expense.getId(),
+                        expense.getPaidBy().getName(),
+                        participant.getEvent().getCode()
+                );
+            }
 
         debtController.generateDebts(participant.getEvent().getCode());
     }
